task: split conditions and clean code (#10)
