+++ src/main/java/org/trustify/operator/cdrs/v2alpha1/server/db/deployment/DBDeploymentReadyPostCondition.java
@@ -1,14 +1,15 @@
-package org.trustify.operator.cdrs.v2alpha1.server.db.utils;
+package org.trustify.operator.cdrs.v2alpha1.server.db.deployment;
 
 import io.fabric8.kubernetes.api.model.apps.Deployment;
 import io.javaoperatorsdk.operator.api.reconciler.Context;
 import io.javaoperatorsdk.operator.api.reconciler.dependent.DependentResource;
+import io.javaoperatorsdk.operator.processing.dependent.workflow.Condition;
 import org.trustify.operator.cdrs.v2alpha1.Trustify;
-import org.trustify.operator.cdrs.v2alpha1.server.db.deployment.DBDeploymentDiscriminator;
 
-public class DBUtils {
+public class DBDeploymentReadyPostCondition implements Condition<Deployment, Trustify> {
 
-    public static boolean isDeploymentReady(DependentResource<Deployment, Trustify> dependentResource, Trustify primary, Context<Trustify> context) {
+    @Override
+    public boolean isMet(DependentResource<Deployment, Trustify> dependentResource, Trustify cr, Context<Trustify> context) {
         return context.getSecondaryResource(Deployment.class, new DBDeploymentDiscriminator())
                 .map(deployment -> {
                     final var status = deployment.getStatus();
